requirements.txt

# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
asyncpg>=0.29.0

# Authentication & Security
supabase>=2.0.0  # CORRECT AUTH - Supabase replaces custom JWT
python-jose[cryptography]>=3.3.0  # DEPRECATED - Remove after migration
passlib[bcrypt]>=1.7.4  # DEPRECATED - Remove after migration
python-decouple>=3.8
pydantic[email]>=2.0.0
PyJWT>=2.8.0  # DEPRECATED - Remove after migration

# AI & ML
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langsmith>=0.0.87
faiss-cpu>=1.7.4

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# API Clients
stripe>=5.0.0
sendgrid>=6.9.0
google-ads>=22.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Monitoring & Logging
sentry-sdk[fastapi]>=1.32.0
psutil>=5.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0.1
Pillow>=10.0.0
numpy>=1.24.0
psutil==5.9.8
scikit-learn==1.3.0
joblib==1.3.2
